% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{commandEnvironments}
\alias{commandEnvironments}
\title{Read environments from a file list to a single environment
Deprecated name (a trivial wrapper now for loadEnvironments)
FIXME should just be an alias, if I understood rdnames}
\usage{
commandEnvironments(..., parent = parent.frame())
}
\arguments{
\item{...}{parameters to pass to loadEnvironments}

\item{parent}{get and pass on our parent}
}
\description{
Read environments from a file list to a single environment
Deprecated name (a trivial wrapper now for loadEnvironments)
FIXME should just be an alias, if I understood rdnames
}
