% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{csvRead}
\alias{csvRead}
\title{read a csv matched from a list of files}
\usage{
csvRead(
  pat = NULL,
  exts = c("csv", "CSV"),
  fl = makeArgs(),
  readFun = readr::read_csv,
  ...
)
}
\arguments{
\item{pat}{optional string pattern to match}

\item{exts}{extensions for the table file}

\item{fl}{file list (defaults to command arguments)}

\item{readFun}{function to use for reading}

\item{...}{additional parameters for readFun}
}
\description{
read a csv matched from a list of files
}
