% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{csvReadList}
\alias{csvReadList}
\title{read csv files into a list}
\usage{
csvReadList(
  pat = NULL,
  exts = c("csv", "CSV"),
  fl = makeArgs(),
  ...,
  names = NULL,
  trim = "\\\\.[^.]*$"
)
}
\arguments{
\item{pat}{optional string pattern to match}

\item{exts}{extensions for the table file}

\item{fl}{file list (defaults to command arguments)}

\item{...}{additional parameters for read_csv}

\item{names}{for the new list}

\item{trim}{if names are not given, take stuff off the end of the matched filenames (defaults to the last "." and everything thereafter)}
}
\description{
read csv files into a list
}
