% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{csvSave}
\alias{csvSave}
\title{Save a table to csv}
\usage{
csvSave(
  ...,
  target = targetname(),
  ext = "Rout.csv",
  writeFun = readr::write_csv
)
}
\arguments{
\item{...}{arguments to write_csv (typically, name of variable to save)}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (.Rout.csv)}

\item{writeFun}{function for writing}
}
\description{
Save a table to csv
}
