% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{getEnvironment}
\alias{getEnvironment}
\title{get a single environment using matchFile, load it and return it}
\usage{
getEnvironment(pat = "", fl = makeArgs(), exts = c("RData", "rda", "rdata"))
}
\arguments{
\item{pat}{pattern to match}

\item{fl}{file list to select from (makeArgs by default)}

\item{exts}{extensions to select}
}
\description{
get a single environment using matchFile, load it and return it
}
