% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{getEnvObj}
\alias{getEnvObj}
\title{get a single object from a single environment}
\usage{
getEnvObj(oname, pat = "", ...)
}
\arguments{
\item{oname}{object name in the environment}

\item{pat}{pattern to match environment name}

\item{...}{arguments for getEnvironment()}
}
\description{
get a single object from a single environment
}
