% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{loadEnvironmentList}
\alias{loadEnvironmentList}
\title{read environment files and return a list of environments}
\usage{
loadEnvironmentList(
  pat = NULL,
  fl = makeArgs(),
  exts = c("RData", "rda", "rdata"),
  names = NULL,
  trim = "\\\\.[^.]*$"
)
}
\arguments{
\item{pat}{pattern to match}

\item{fl}{file list to select from (makeArgs by default)}

\item{exts}{extensions to select}

\item{names}{a list of names for the environments found}

\item{trim}{a pattern to eliminate when making names from the file names}
}
\description{
read environment files and return a list of environments
}
