% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{loadEnvironments}
\alias{loadEnvironments}
\title{Load a list of environments}
\usage{
loadEnvironments(envl, parent = parent.frame())
}
\arguments{
\item{envl}{a list of environments}

\item{parent}{(defaults to parent.frame())}
}
\description{
Load a list of environments
}
