% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{makeArgs}
\alias{makeArgs}
\title{Not exported: makeArgs
A service function to get the make arguments
when R was called interactively, these come from a variable called callArgs
otherwise parsed from the command line}
\usage{
makeArgs()
}
\description{
Not exported: makeArgs
A service function to get the make arguments
when R was called interactively, these come from a variable called callArgs
otherwise parsed from the command line
}
