% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{makeGraphics}
\alias{makeGraphics}
\title{Open a graphics file for saving (internal function; exported for legacy reasons; use startGraphics)}
\usage{
makeGraphics(
  target = makeArgs()[[1]],
  otype = NULL,
  ext = otype,
  desc = NULL,
  ...
)
}
\arguments{
\item{target}{stem of filename (defaults to the target of the script)}

\item{otype}{type of file to write (will use pdf if not specified)}

\item{ext}{file extension (will use pdf.tmp if not specified)}

\item{desc}{text to replace Rout in default target}

\item{...}{arguments to pass to grDevices call}
}
\description{
Open a graphics file for saving (internal function; exported for legacy reasons; use startGraphics)
}
