% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{matchFile}
\alias{matchFile}
\title{Try to find a single file matching a given description (usually for reading)}
\usage{
matchFile(pat, fl = makeArgs(), exts = NULL)
}
\arguments{
\item{pat}{is a pattern for matching}

\item{fl}{is the list of filenames to search (makeArgs by default)}

\item{exts}{is a list of allowed extensions}
}
\description{
Try to find a single file matching a given description (usually for reading)
}
