% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{outFile}
\alias{outFile}
\title{Write to a different kind of file}
\usage{
outFile(..., target = targetname(), ext = "Rout.txt")
}
\arguments{
\item{...}{arguments to sync}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (.Rout.txt)}
}
\description{
Write to a different kind of file
}
