% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pipeStar}
\alias{pipeStar}
\title{Return the starred text passed by make}
\usage{
pipeStar(tag = "pipestar", fl = makeArgs())
}
\arguments{
\item{tag}{identifies the passed text}

\item{fl}{is the list of filenames to search (makeArgs by default)}
}
\description{
Return the starred text passed by make
}
