% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{rdaSave}
\alias{rdaSave}
\title{Save variables to rdata
Convenience wrapper for saveEnvironment (all variables) or saveVars (specific variables)}
\usage{
rdaSave(..., target = targetname(), ext = "rda")
}
\arguments{
\item{...}{names of variables to save, defaults to NULL which is interpreted as all}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (rda)}
}
\description{
Save variables to rdata
Convenience wrapper for saveEnvironment (all variables) or saveVars (specific variables)
}
