% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rdsRead}
\alias{rdsRead}
\title{Read an rds file matched from a list of files}
\usage{
rdsRead(
  pat = NULL,
  exts = c("rds", "Rds", "RDS"),
  fl = makeArgs(),
  refhook = NULL
)
}
\arguments{
\item{pat}{optional string pattern to match}

\item{exts}{extensions for the table file}

\item{fl}{file list (defaults to command arguments)}

\item{refhook}{pass to readRDS}
}
\description{
Read an rds file matched from a list of files
}
