% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rdsReadList}
\alias{rdsReadList}
\title{read rds files and return a list of objects}
\usage{
rdsReadList(
  pat = NULL,
  fl = makeArgs(),
  exts = c("rds", "RDS"),
  names = NULL,
  trim = "\\\\.[^.]*$",
  refhook = NULL
)
}
\arguments{
\item{pat}{pattern to match}

\item{fl}{file list to select from (makeArgs by default)}

\item{exts}{extensions to select}

\item{names}{a list of names for the environments found}

\item{trim}{a pattern for making names from the file names (not used when names if provided)}

\item{refhook}{pass to readRDS}
}
\description{
read rds files and return a list of objects
}
