% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{rdsSave}
\alias{rdsSave}
\title{Save a variable to rds}
\usage{
rdsSave(vname, target = targetname(), ext = "rds")
}
\arguments{
\item{vname}{name of variable}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (rds)}
}
\description{
Save a variable to rds
}
