% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{rpcall}
\alias{rpcall}
\title{set callArgs for interactive use}
\usage{
rpcall(call)
}
\arguments{
\item{call}{command-line call as generated in .Rout.args}
}
\description{
set callArgs for interactive use
}
