% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveEnvironment}
\alias{saveEnvironment}
\title{Save the whole environment}
\usage{
saveEnvironment(target = targetname(), ext = "rda")
}
\arguments{
\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (rda)}
}
\description{
Save the whole environment
}
