% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveVars}
\alias{saveVars}
\title{Save particular variables to rdata}
\usage{
saveVars(..., target = targetname(), ext = "rda")
}
\arguments{
\item{...}{arguments to save (typically, names of variables to save)}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (rda)}
}
\description{
Save particular variables to rdata
}
