% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{soLoad}
\alias{soLoad}
\title{load a TMB shared object}
\usage{
soLoad(pat = NULL, exts = c("so", "SO"), fl = makeArgs())
}
\arguments{
\item{pat}{optional string pattern to match}

\item{exts}{extensions for the table file}

\item{fl}{file list (defaults to command arguments)}
}
\description{
load a TMB shared object
}
