% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tableRead}
\alias{tableRead}
\title{read a table matched from a list of files}
\usage{
tableRead(
  pat = NULL,
  delim = " ",
  exts = c("csv", "CSV", "ssv", "scsv", "tsv"),
  fl = makeArgs(),
  ...
)
}
\arguments{
\item{pat}{optional string pattern to match}

\item{delim}{delimiter for reading files}

\item{exts}{extensions for the table file}

\item{fl}{file list (defaults to command arguments)}

\item{...}{additional parameters for read_delim}
}
\description{
read a table matched from a list of files
}
