% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{targetname}
\alias{targetname}
\title{Get a targetname}
\usage{
targetname(ext = "", suffix = "\\\\.Rout", fn = makeArgs()[[1]])
}
\arguments{
\item{ext}{file extension for output}

\item{suffix}{file extension of provided name (.Rout by default)}

\item{fn}{provided file name (first of commandArgs by default)}
}
\description{
Get a targetname
}
