% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{teeGG}
\alias{teeGG}
\title{saveGG and also print to stdout(for quick reference)}
\usage{
teeGG(
  g,
  target = makeArgs()[[1]],
  ext = "pdf",
  desc = "ggp",
  print_title = desc,
  crop = TRUE,
  ...
)
}
\arguments{
\item{g}{graphical object to print and save}

\item{target}{stem of filename (defaults to the target of the script)}

\item{ext}{file extension (will use pdf if not specified)}

\item{desc}{text to replace Rout in default target (ggp by default)}

\item{print_title}{plot title for the printed version (defaults to desc part of filename)}

\item{crop}{logical: use knitr::plot_crop?}

\item{...}{arguments to pass to ggsave call}
}
\description{
saveGG and also print to stdout(for quick reference)
}
