% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tsvRead}
\alias{tsvRead}
\title{read a tsv matched from a list of files}
\usage{
tsvRead(pat = NULL, exts = c("tsv", "TSV"), fl = makeArgs(), ...)
}
\arguments{
\item{pat}{optional string pattern to match}

\item{exts}{extensions for the table file}

\item{fl}{file list (defaults to command arguments)}

\item{...}{additional parameters for read_tsv}
}
\description{
read a tsv matched from a list of files
}
