% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{tsvSave}
\alias{tsvSave}
\title{Save a table to tsv}
\usage{
tsvSave(..., target = targetname(), ext = "Rout.tsv")
}
\arguments{
\item{...}{arguments to write_tsv (typically, name of variable to save)}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (.Rout.tsv)}
}
\description{
Save a table to tsv
}
