% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{fileSelect}
\alias{fileSelect}
\title{Not exported: fileSelect
Select a list of filenames from a list by extension
… and optionally by pattern as well
not currently exported; used by other file selectors}
\usage{
fileSelect(fl = makeArgs(), exts = NULL, pat = NULL)
}
\arguments{
\item{fl}{filelist (makeArgs by default)}

\item{exts}{a list of extensions to match (. is added here)}

\item{pat}{a pattern to filter by}
}
\description{
Not exported: fileSelect
Select a list of filenames from a list by extension
… and optionally by pattern as well
not currently exported; used by other file selectors
}
