% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{rdsSave}
\alias{rdsSave}
\title{Serialize with saveRDS using a target stem}
\usage{
rdsSave(object, target = targetname(), ext = "rds", printSummary = FALSE)
}
\arguments{
\item{object}{R object to save}

\item{target}{stem of file to save to (defaults to name from call)}

\item{ext}{file extension (rds)}

\item{printSummary}{logical}
}
\description{
Serialize with saveRDS using a target stem
}
