% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{saveGG}
\alias{saveGG}
\title{Apparently ggplot doesn't support pdf well ☹}
\usage{
saveGG(g, target = makeArgs()[[1]], ext = "pdf", desc = "ggp", ...)
}
\arguments{
\item{g}{graphical object to print}

\item{target}{stem of filename (defaults to the target of the script)}

\item{ext}{file extension (will use pdf if not specified)}

\item{desc}{text to replace Rout in default target (ggp by default)}

\item{...}{arguments to pass to ggsave call}
}
\description{
Apparently ggplot doesn't support pdf well ☹
}
