% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{saveGG}
\alias{saveGG}
\title{save a ggplot object with a name derived from the target name
this is good because ggplot objects don't always play well with print
and because it's good to attach mnemonic names when you have a lot of files
TODO: Should probably have logic for mnemonic names for base plots}
\usage{
saveGG(
  g,
  target = makeArgs()[[1]],
  ext = "pdf",
  desc = "ggp",
  crop = TRUE,
  ...
)
}
\arguments{
\item{g}{graphical object to print}

\item{target}{stem of filename (defaults to the target of the script)}

\item{ext}{file extension (will use pdf if not specified)}

\item{desc}{text to replace Rout in default target (ggp by default)}

\item{crop}{logical: use knitr::plot_crop?}

\item{...}{arguments to pass to ggsave call}
}
\description{
save a ggplot object with a name derived from the target name
this is good because ggplot objects don't always play well with print
and because it's good to attach mnemonic names when you have a lot of files
TODO: Should probably have logic for mnemonic names for base plots
}
