% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{sourceFiles}
\alias{sourceFiles}
\title{Source certain files from a file list
Usually we are inside the first .R file so we want to source the others (first=FALSE by default)
If called from a wrapper, we probably want to source the main source file (first in list) _last_.
Other styles of calling need to be done in a clunky, manual fashion; I've never needed to do this}
\usage{
sourceFiles(
  fl = makeArgs(),
  exts = c("R", "r"),
  first = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{fl}{file list to select from (makeArgs by default)}

\item{exts}{extensions to select}

\item{first}{(Boolean); should we source the first matched file (usually the master script)?}

\item{verbose}{(Boolean); passed to source()}
}
\description{
Source certain files from a file list
Usually we are inside the first .R file so we want to source the others (first=FALSE by default)
If called from a wrapper, we probably want to source the main source file (first in list) _last_.
Other styles of calling need to be done in a clunky, manual fashion; I've never needed to do this
}
