% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{startGraphics}
\alias{startGraphics}
\title{Open a graphics file for saving}
\usage{
startGraphics(
  ...,
  target = makeArgs()[[1]],
  otype = "pdf",
  ext = otype,
  always = FALSE
)
}
\arguments{
\item{...}{arguments to pass to grDevices call (via makeGraphics)}

\item{target}{stem of filename (defaults to the target of the script)}

\item{otype}{type of file to write (defaults to pdf)}

\item{ext}{file extension (defaults to otype)}

\item{always}{open file even if called from interactive model}
}
\description{
Open a graphics file for saving
}
